fix : removed percent time in range as it is not calculating over time.
