+++ .pre-commit-config.yaml
@@ -1,0 +1,20 @@
+repos:
+- repo: https://github.com/pre-commit/pre-commit-hooks
+  rev: v2.3.0
+  hooks:
+      - id: check-yaml
+      - id: end-of-file-fixer
+      - id: trailing-whitespace
+-   repo: https://github.com/ambv/black
+    rev: 19.10b0
+    hooks:
+    -   id: black
+        args: [--line-length=120]
+-   repo: https://gitlab.com/pycqa/flake8
+    rev: 3.7.9
+    hooks:
+    -   id: flake8
+        args: [
+          '--max-line-length', '120',
+          '--ignore', 'E203,E266,E501,W503,F401,F403'
+        ]
