+++ requirements.txt
@@ -1,4 +1,5 @@
 black==19.10b0
+flake8==3.7.9
 numpy==1.18.1
 pandas==1.0.3
 pytest==5.4.1
