+++ tests/common/test_common.py
@@ -1,4 +1,5 @@
-from tidepool_data_science_metrics.common.common import mean, avg, std_deviation
+from tidepool_data_science_metrics.common.common \
+    import mean, avg, std_deviation
 
 
 def test_avg_glucose(bg_values):
