+++ tests/conftest.py
@@ -1,5 +1,6 @@
 import pytest
 import numpy as np
+import datetime
 
 
 @pytest.fixture
@@ -107,4 +108,57 @@ def bg_values(scope="module"):
             [60],
             [50],
         ]
+    )
+
+
+@pytest.fixture
+def get_date_values(scope="module"):
+    return np.array(
+        [
+            ["8/15/2019 00:40:00", 87],
+            ["8/15/2019 00:45:00", 90],
+            ["8/15/2019 01:55:00", 104],
+            ["8/15/2019 00:50:00", 92],
+            ["8/15/2019 00:55:00", 96],
+            ["8/15/2019 01:00:00", 94],
+            ["8/15/2019 01:05:00", 97],
+            ["8/15/2019 01:10:00", 95],
+            ["8/15/2019 01:15:00", 100],
+            ["8/15/2019 01:20:00", 102],
+            ["8/15/2019 01:25:00", 101],
+            ["8/15/2019 01:30:00", 105],
+            ["8/15/2019 01:35:00", 80],
+            ["8/15/2019 01:40:00", 103],
+            ["8/15/2019 01:45:00", 108],
+            ["8/15/2019 01:50:00", 103],
+            ["8/16/2019 02:00:00", 108],
+        ],
+        dtype=object,
+    )
+
+
+@pytest.fixture
+def get_date_ep_values(scope="module"):
+    return np.array(
+        [
+            ["8/15/2019 00:40:00", 87],
+            ["8/15/2019 00:45:00", 90],
+            ["8/15/2019 00:50:00", 44],
+            ["8/15/2019 00:55:00", 46],
+            ["8/15/2019 01:00:00", 51],
+            ["8/15/2019 01:05:00", 97],
+            ["8/15/2019 01:10:00", 95],
+            ["8/15/2019 01:15:00", 100],
+            ["8/15/2019 01:20:00", 54],
+            ["8/15/2019 01:25:00", 101],
+            ["8/15/2019 01:30:00", 105],
+            ["8/15/2019 01:35:00", 44],
+            ["8/15/2019 01:40:00", 80],
+            ["8/15/2019 01:45:00", 40],
+            ["8/15/2019 01:50:00", 50],
+            ["8/15/2019 01:55:00", 52],
+            ["8/16/2019 02:00:00", 45],
+            ["8/16/2019 02:05:00", 108],
+        ],
+        dtype=object,
     )
