+++ tidepool_data_science_metrics/__init__.py
@@ -1,19 +1,2 @@
 from tidepool_data_science_metrics.cgm import cgm
 from tidepool_data_science_metrics.common import common
-
-"""   
-from tidepool_data_science_metrics.cgm2 import (
-    cgm3.get_percent_values_by_range,
-    get_percent_time_in_range,
-    get_cv_of_glucose,
-    get_gmi,
-    get_bgri,
-    get_episodes,
-)
-
-from tidepool_data_science_metrics.common.common import (
-    get_mean_glucose,
-    get_avg_glucose,
-    get_std_deviation
-)
-"""
