+++ tox.ini
@@ -1,0 +1,5 @@
+[tool.black]
+line-length = 120
+
+[flake8]
+max-line-length = 120
